Rename steps -> steppy
