+++ steppy/misc.py
@@ -2,8 +2,8 @@ from attrdict import AttrDict
 import lightgbm as lgb
 from sklearn.externals import joblib
 
-from steps.base import BaseTransformer
-from steps.utils import get_logger
+from steppy.base import BaseTransformer
+from steppy.utils import get_logger
 
 logger = get_logger()
