+++ steppy/postprocessing.py
@@ -3,7 +3,7 @@ import numpy as np
 import pandas as pd
 from sklearn.externals import joblib
 
-from .base import BaseTransformer
+from steppy.base import BaseTransformer
 
 
 class ClassPredictor(BaseTransformer):
