+++ tests/test_base.py
@@ -1,21 +1,7 @@
 import pytest
-
-from steps.base import make_transformer
-
-
-@pytest.mark.parametrize("mode", [0, 1])
-def test_make_transformer(mode):
-    def fun(x, y, mode=0):
-        return x + y if mode == 0 else x - y
-    tr = make_transformer(fun)
-
-    tr.fit()
-    res = tr.transform(7, 3, mode=mode)
-    assert res == (10 if mode == 0 else 4)
-
 import numpy as np
 
-from steps.base import Step, NoOperation, StepsError
+from steppy.base import Step, NoOperation, StepsError, make_transformer
 
 from .steps_test_utils import CACHE_DIRPATH
 
@@ -47,6 +33,17 @@ def data():
             'labels': np.array([1, 1, 0])
         }
     }
+
+
+@pytest.mark.parametrize("mode", [0, 1])
+def test_make_transformer(mode):
+    def fun(x, y, mode=0):
+        return x + y if mode == 0 else x - y
+    tr = make_transformer(fun)
+
+    tr.fit()
+    res = tr.transform(7, 3, mode=mode)
+    assert res == (10 if mode == 0 else 4)
 
 
 def test_inputs_without_conflicting_names_do_not_require_adapter(data):
